Moved now unused queue based async into new file
Failed experiment backup, might salvage some things out of it later on.
